moved macros out to their own module, added an initial test
